Rename EditStatusButton to generic PopoverButton
